refactor(frontend): isolate chat cancellation helper from sse adapter (#460)
